ai_api/requirements.txt
openai==1.53.0
python-dotenv==1.0.1
fastapi==0.115.4